수정내역 1. Player의 MoveAction을 조금 더 디테일하게 수정 2. Player의 Passive 수치 조작 처리

+++ Framework/Manager/CharacterManager.h
@@ -27,6 +27,6 @@ struct CharacterManager
 	int characterHp;
 	int characterAttack;
 	int characterDefence;
-	int characterSpeed;
+	float characterSpeed;
 	int Gold;
 };
